Django/ShibService/requirements.txt
Django == 2.1.3
django-bootstrap-static==4.0.0
django-qr-code==1.0.0